CO-447 Move contrib/pkg/installmanager/ to pkg/installmanager
