+++ pkg/installmanager/installmanager.go
