+++ pkg/installmanager/installmanager_test.go
